md partial: simplify status processing
Ignore the status of an entry entirely. Just sync each mentioned
(section, name) pair once. For entries that are later aborted or
otherwise don't complete, this causes a couple extra metadata
requests, but these are inexpensive and should be rare. This is
guaranteed to catch all updates, and is much simpler to understand and
verify. Data sync may require more complex handling, but metadata sync
is cheap enough to be able to use a simpler approach.

Signed-off-by: Josh Durgin <josh.durgin@inktank.com>