client: clean up request generation
Instead of having special strings as commands and handling those
in a bunch of conditionals, put the path and data required directly
in the method that uses it. This makes it easier to see the correspondence
between a method and the API that it's calling.

Add a general connection method so other files don't duplicate code.

Signed-off-by: Josh Durgin <josh.durgin@inktank.com>