packaging: rm unnecessary RPM tags
The BuildRoot tag is only needed on RHEL 5. RHEL 6 and newer are fine
without this setting.

The %{_prefix} macro is already /usr by default.

The Vendor tag is something that should be left undefined in the
packaging .spec itself. Ordinarily this is added during the build by the
build system.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>