radosgw-sync: keep full-sync state
Use a new replicalog construct to keep track of the full sync stage.
Set state to 'init'. Generate a list of all the buckets that need to exist, update all
these buckets as keys in the replicalog. Then set the state to
'full-sync'. Later we'll set the state to 'incremental'.

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>